Adicionado função para carregamento de fotos
